# packages/api-schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nature Explorer API",
    "version": "0.1.0",
    "description": "Generated at startup from the TypeBox schemas of the registered routes. Every non-2xx response uses the shared `Error` schema and echoes the request id in `x-request-id`. `/v1/me*` and `/v1/auth/logout` need a bearer access token (`bearerAuth`)."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /v1/auth/apple or /v1/auth/refresh (HS256, 15 minutes)"
      }
    },
    "schemas": {
      "Error": {
        "additionalProperties": false,
        "description": "Shared error envelope returned by every non-2xx response. `code` is a stable machine-readable identifier in SCREAMING_SNAKE_CASE; `message` is for developers, never shown verbatim to players; `details` carries validation issues or other structured context.",
        "type": "object",
        "required": ["error", "requestId"],
        "properties": {
          "error": {
            "additionalProperties": false,
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "pattern": "^[A-Z][A-Z0-9_]+$",
                "type": "string",
                "example": "VALIDATION_FAILED"
              },
              "message": { "type": "string" },
              "details": { "additionalProperties": true, "type": "object", "properties": {} }
            }
          },
          "requestId": {
            "description": "Fastify request id, also returned in the `x-request-id` response header",
            "type": "string"
          }
        }
      },
...
      "Me": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "displayName",
          "factionId",
          "factionChangedAt",
          "factionChangeAvailableAt",
          "xp",
          "level",
          "role",
          "createdAt",
          "suggestedFactionId"
        ],
        "properties": {
          "id": { "format": "uuid", "type": "string" },
          "displayName": { "type": "string" },
          "factionId": { "type": ["null", "integer"] },
          "factionChangedAt": { "format": "date-time", "type": ["null", "string"] },
          "factionChangeAvailableAt": {
            "format": "date-time",
            "description": "When the next faction change is allowed; null when a change is allowed now (or no faction is set yet)",
            "type": ["null", "string"]
          },
          "xp": { "minimum": 0, "type": "integer" },
          "level": { "minimum": 1, "type": "integer" },
          "role": { "type": "string", "enum": ["player", "tester", "admin"] },
          "createdAt": { "format": "date-time", "type": "string" },
          "suggestedFactionId": {
            "description": "Faction with the fewest active members (tie → lowest id)",
            "type": "integer"
          }
        }
      },
      "MeUpdate": {
        "additionalProperties": false,
        "type": "object",
        "required": ["displayName"],
        "properties": {
          "displayName": {
            "minLength": 1,
            "maxLength": 64,
            "description": "Validated by the service to 2–24 Unicode scalars after trimming, no control characters",
            "type": "string"
          }
        }
      },
      "FactionSelect": {
        "additionalProperties": false,
        "type": "object",
        "required": ["factionId"],
        "properties": { "factionId": { "minimum": 1, "maximum": 32767, "type": "integer" } }
      },
      "AccountDeletion": {
        "additionalProperties": false,
        "type": "object",
        "required": ["deletedAt", "purgeAt"],
        "properties": {
          "deletedAt": { "format": "date-time", "type": "string" },
          "purgeAt": {
            "format": "date-time",
            "description": "When the erasure job runs (deletedAt + 30 days)",
            "type": "string"
          }
        }
      },
      "ExportStatus": {
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "status",
          "requestedAt",
          "completedAt",
          "expiresAt",
          "downloadUrl",
          "error"
        ],
        "properties": {
          "id": { "format": "uuid", "type": "string" },
          "status": { "type": "string", "enum": ["pending", "ready", "failed"] },
          "requestedAt": { "format": "date-time", "type": "string" },
          "completedAt": { "format": "date-time", "type": ["null", "string"] },
          "expiresAt": { "format": "date-time", "type": ["null", "string"] },
          "downloadUrl": {
            "description": "Presigned GET URL, valid for one hour; only when status is ready",
            "type": ["null", "string"]
          },
          "error": { "type": ["null", "string"] }
        }
      },
...
    }
  },
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "getMe",
        "summary": "The signed-in player's profile",
        "tags": ["me"],
        "description": "Includes `suggestedFactionId` so the faction pick screen needs no second call. Never returns the e-mail address.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Profile",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Me" } } }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateMe",
        "summary": "Update the display name",
        "tags": ["me"],
        "description": "The display name is trimmed, must be 2–24 Unicode scalars long and contain no control characters. Violations answer 400 VALIDATION_FAILED with `details.field = \"displayName\"` and `details.rule` in tooShort | tooLong | controlCharacter.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/MeUpdate" } }
          }
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Me" } } }
          },
          "400": {
            "description": "Request validation failed (code VALIDATION_FAILED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMe",
        "summary": "Delete the account (30-day grace period)",
        "tags": ["me"],
        "description": "Marks the account deleted, revokes every refresh token and schedules erasure 30 days later (`account.purge`). Signing in with the same Apple ID before `purgeAt` restores the account. After this call every request with the old credentials answers 401 ACCOUNT_DELETED.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Deletion accepted",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AccountDeletion" } }
            }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/v1/me/faction": {
      "post": {
        "operationId": "selectFaction",
        "summary": "Pick or change the faction",
        "tags": ["me"],
        "description": "The first pick is free and starts no lock. Re-selecting the current faction is a no-op. A change is allowed when the player has never changed or the previous change is at least 30 days old; otherwise 409 FACTION_CHANGE_LOCKED with `details.nextChangeAt`. XP and level are never altered.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/FactionSelect" } }
          }
        },
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Updated profile (with `factionChangeAvailableAt` and `suggestedFactionId`)",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Me" } } }
          },
          "400": {
            "description": "Request validation failed (code VALIDATION_FAILED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "404": {
            "description": "Unknown faction id (code FACTION_NOT_FOUND)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          },
          "409": {
            "description": "Faction change locked (code FACTION_CHANGE_LOCKED, `details.nextChangeAt` ISO 8601)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    },
    "/v1/me/export": {
      "get": {
        "operationId": "getExport",
        "summary": "Request or poll the player's data export",
        "tags": ["me"],
        "description": "Returns the latest export when it is pending, ready and not older than 7 days (with a fresh `downloadUrl` valid for one hour), or failed less than a minute ago (so the failure is visible). Otherwise enqueues a new export (`account.export`) and answers 202 with `status: pending`. Clients poll every 5 seconds while pending.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Existing export (pending, ready or failed)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExportStatus" } }
            }
          },
          "202": {
            "description": "A new export was enqueued",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExportStatus" } }
            }
          },
          "401": {
            "description": "Missing or invalid access token (UNAUTHORIZED), expired token (TOKEN_EXPIRED) or deleted account (ACCOUNT_DELETED)",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Error" } }
            }
          }
        }
      }
    }
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local development (make dev)" }],
...
}